💄 start on styling--issues with hmr
