+++ apps/edge/recoverage-server/__tests__/index.test.ts
@@ -1,6 +1,6 @@
 import { env } from "cloudflare:test"
 
-import app from "../src/app"
+import app from "../src"
 
 console.log(env)
